+++ .editorconfig
@@ -22,3 +22,5 @@ insert_final_newline = true
 [*.{js,txt,md,css,html,php,py,json,yml,sass,pug}]
 indent_style = space
 indent_size = 2
+insert_final_newline = true
+
