+++ js/variables.js
@@ -1,0 +1,29 @@
+const MESSAGES = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = ['Оля', 'Петя', 'Коля', 'Саша', 'Катя', 'Ира'];
+
+const DESCRIPTIONS = [
+  'Приятно провожу вечер',
+  'Скорее бы выходные!',
+  'Очень много работы',
+  'Просто хорошее настроение',
+  'Отдыхаю с друзьями',
+  'Инстаграм для бумеров',
+];
+
+const USERS_POSTS_COUNTS = 25;
+const MIN_AVATAR = 1;
+const MAX_AVATAR = 6;
+const MIN_INDEX_FOTO = 1;
+const MAX_INDEX_FOTO = 25;
+const MIN_LIKE = 15;
+const MAX_LIKE = 200;
+const USERS_COMMENTS_COUNTS = 5;
+
+export {MIN_AVATAR, MAX_AVATAR, MESSAGES, NAMES, USERS_COMMENTS_COUNTS, MIN_INDEX_FOTO, MAX_INDEX_FOTO, DESCRIPTIONS, MIN_LIKE, MAX_LIKE, USERS_POSTS_COUNTS};
